requirements.txt
Django==5.0.2
django-allauth==0.61.1
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
django-colorfield==0.11.0
django-taggit==5.0.1
django-ckeditor==6.7.0
gunicorn==21.2.0
whitenoise==6.6.0
python-dotenv==1.0.1 